chore: 建立 pyproject.toml 与 ruff 配置，生产代码 lint 清零
依赖按用途分组：核心出片链路只需 numpy，重依赖按需引入。
- audio: librosa 必须约束 numba>=0.61，否则会解析到不兼容 Python 3.12
  的 numba 0.53 并编译失败
- separate: audio-separator 不会自动带上 onnxruntime，必须显式声明

ruff 配置的两处项目特有决定：
- 忽略 RUF001/002/003：本项目注释、文档与错误信息一律用中文，
  全角标点是有意为之，这三条在中文项目里全是误报（实测 319 条告警
  几乎全部来自此）
- 忽略 S603/S607：大量调用 ffmpeg/yt-dlp，命令由代码构造、不拼接用户输入

修复：
- 删除 ass_builder 重构后遗留的死变量 line_start
- zip 显式 strict=False：各处长度理应相等，但渲染路径不应因此抛异常中断出片
- 信用窗口扫描改用 itertools.pairwise

回归验证：改动后重新生成 ASS，仍为 451 个 Dialogue 事件，与改动前一致。

+++ backend/kvm/pipeline/beat_detect.py
@@ -25,7 +25,7 @@ class BeatGrid:
     """每个拍点的时间（毫秒），升序。"""
 
     def beat_period_ms(self) -> int:
-        return int(round(60000.0 / self.bpm)) if self.bpm > 0 else 0
+        return round(60000.0 / self.bpm) if self.bpm > 0 else 0
 
     def beats_before(self, t_ms: int, n: int) -> list[int]:
         """返回 `t_ms` 之前最近的 n 个拍点，升序。不足 n 个则有多少给多少。"""
@@ -52,7 +52,7 @@ def detect_beats(audio_path: Path, *, drums_path: Path | None = None) -> BeatGri
         if times is None or len(times) < 4:
             return None
         bpm = float(np.atleast_1d(tempo)[0])
-        return BeatGrid(bpm=bpm, beats_ms=[int(round(t * 1000)) for t in times])
+        return BeatGrid(bpm=bpm, beats_ms=[round(t * 1000) for t in times])
     except Exception:
         # 节拍检测是增强项，任何失败都不应阻断出片
         return None
